Fix numpy usage on the master node
Note that slave nodes will not be able to use numpy since the makefile is only executed on the master node. The numpy wheel was compiled with this dependency available, it is unclear if we can compile it without this system package dependency. If we figure out a way to do this, we can re-compile the wheel and revert this change.

Change-Id: I01bbc96aecd2fc0bf08c0466550033d0db15aacd