Upgrade edx-enterprise-data to 0.2.5 and unpin requirements
Unpinned a number of requirements so they would be upgraded as part of
`make upgrade` as recommended by OEP-18. In particular the edx-enterprise-data
django app was not pinning edx-drf-extensions which resulted in 2 different
versions of DRF in the requirements (one for enterprise-data and one for this
IDA).

Upgrading DRF resulting in an assertion failure that was added in DRF 3.7.4
```
Cannot both declare the field 'created' and include it in the CourseProgramMetadataSerializer
'exclude' option. Remove the field or, if inherited from a parent serializer, disable with `created = None`.
```

https://github.com/encode/django-rest-framework/pull/5599

Because the created date was being excluded, it is not needed in the CourseProgramMetadataSerializer and
has been removed.