feat: Make v1 of Analytics Data API Devstack Testable
This change makes v1 of the Analytics Data API testable in devstack.

The changes include:
* modifying the generate_fake_course_data management command to accept a "database" option
* modifying the loaddata make target to run the generate_fake_course_data management command with the analytics_v1 database
* adding a new problem_response_answer_distribution_analytics_v1 JSON file and loading it as part of the loaddata make target
* adding a development specific Django router
* modifying Django settings in the based, development, and test environment to use the correct Django routers

When users of devstack run the loaddata make target, both the analytics and the analytics_v1 databases will be loaded with dummy data. The data will be different, so users will be able to verify differences in the APIs.