fix: change github_ci Make target to use analytics_v1 database
This pull request changes the github_ci Make target to run the generate_fake_course_data management command with the analytics_v1 option instead of the default, which is the default database. Because we do not migrate v0 app tables into the default database during CI, the tables that the generate_fake_course_data management command tries to write into do not exist.